Fixed naming of the file for future pull request
