+++ CMP1903M A01 2223/CMP1903M A01 2223/Program.cs
@@ -1,0 +1,11 @@
+﻿namespace CMP1903M_A01_2223
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Testing testing = new Testing();
+            
+        }
+    }
+}
