+++ CMP1903M A01 2223/CMP1903M A01 2223/bin/Debug/net6.0/CMP1903M A01 2223.exe
