--- ModifiedCode/CMP1903M A01 2223.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33424.131
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CMP1903M A01 2223", "CMP1903M A01 2223\CMP1903M A01 2223.csproj", "{4D0EBB22-07F5-46C9-970A-C1869E11E4F1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4D0EBB22-07F5-46C9-970A-C1869E11E4F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4D0EBB22-07F5-46C9-970A-C1869E11E4F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4D0EBB22-07F5-46C9-970A-C1869E11E4F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4D0EBB22-07F5-46C9-970A-C1869E11E4F1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F459E5AD-9DB7-44A8-917D-BF473992A227}
-	EndGlobalSection
-EndGlobal
