--- ModifiedCode/CMP1903M A01 2223/bin/Debug/net6.0/CMP1903M A01 2223.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v6.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v6.0": {
-      "CMP1903M A01 2223/1.0.0": {
-        "runtime": {
-          "CMP1903M A01 2223.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "CMP1903M A01 2223/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
