--- ModifiedCode/CMP1903M A01 2223/bin/Debug/net6.0/CMP1903M A01 2223.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net6.0",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "6.0.0"
-    }
-  }
-}
